refactor Dialogue for readings into its own component with state
